refactor: split menu builder from controller
